split timeline_track into small files
